requirements.txt
web3
requests
python-dotenv